requirements.txt
fastapi==0.104.1
pydantic==2.5.1
uvicorn>=0.15.0,<0.16.0
fastapi-sqlalchemy==0.2.1
pydantic-settings==2.1.0
fastapi-sqlalchemy==0.2.1